Fixes context-handling bugs
Todo.txt Touch recognized email as context: fixed
Add task appended context immediately at end of task with a space at the end.
Adds appended context with a space in front as well as after.
Context must have a newline or whitespace in front of the at sign.
Test added for context at start of task.
Closes #19