More robust handling of context and project names.
Both context and project names only matched "word" characters. Basically
alphanumerics and underscore. There are often times where some internal
punction makes sense. (A project of todo.txt or a context of 2010-01-30)

This changes both expressions to match non-space characters, with the
final character being a word character. I added the last to not include
a potential final '.' or ',' if someone were adding the contexts or
projects by hand.