Re-enable sync indicator when pulling. Fixes #282
Fixed a couple of display issues with the taskage field on the task list.
When views are reused, it would sometimes show a date for tasks that had none.
Also, it would show dates even when the Date new tasks preference was not set
(if it had never been set).
Then, I re-enabled calling updateSyncUI when starting a pull. This was
commented out in 180e8bd because it supposedly messed up the current scroll position.
I couldn't reproduce that problem, but did notice the issues noted above. Regardless,
there is no reason to mess with the list when we just want to make the sync icon spin,
so I modified the method to take a boolean parameter. The list is redrawn only when the
parameter is true, and the only time we pass true to the method is after downloading a new
todo.txt file.