Fix for crash after filter.
SwipeDismissList was crashing after getting a touch down event immediately after filtering. It only happened when the list had been scrolled to the bottom before filtering and the the filtered list is too short to fill the screen. I'm not entirely sure why it was happening, but it looks like SwipeDismissList was not getting an UP touch event, so it was not resetting its state. When it later got the DOWN event, it tried to reuse a no longer existent view, and crashed.

I fixed it by resetting the state at the beginning of the DOWN event handler as well as at the end of the UP handler.