Fix relative dates
Simplified the RelativeDate logic to the following:

if date is in future or more than 365 days in past, just display the date in yyyy-mm-dd format.

else if date is more than 30 days in the past, display 'N months ago', where 1 month is considered 30 days. This isn't very accurate, but it should be good enough for our purpose here.

else if date is more than 1 day in the past, display 'N days ago'

else display 'today'.